requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
arabic-reshaper==2.1.4
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.1.2
blinker==1.9.0
cachetools==5.5.2
captcha==0.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
extra-streamlit-components==0.1.80
frozenlist==1.5.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
multidict==6.4.3
narwhals==1.34.1
numpy==1.26.4
packaging==24.2
pandas==2.2.1
pillow==11.2.1
plotly==5.20.0
propcache==0.3.1
protobuf==5.29.4
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
PyJWT==2.10.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.31.0
rpds-py==0.24.0
scipy==1.13.0
six==1.17.0
smmap==5.0.2
streamlit==1.44.1
streamlit-authenticator==0.2.3
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
twilio==9.0.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
yarl==1.19.0
scikit-learn==1.3.2 
